+++ src/Catalog/CompanyListItem.php
@@ -26,6 +26,9 @@ final class CompanyListItem
     /** @var string */
     private $name;
 
+    /** @var null|string */
+    private $corporateName;
+
     /** @var string */
     private $description;
 
@@ -62,6 +65,7 @@ final class CompanyListItem
     {
         $this->id = to_int($data['id']);
         $this->name = to_string($data['name']);
+        $this->corporateName = (\array_key_exists('corporateName', $data) === true) ? $data['corporateName'] : null;
         $this->description = to_string($data['description']);
         $this->address = to_string($data['address']);
         $this->phoneNumber = to_string($data['phoneNumber']);
@@ -99,6 +103,12 @@ final class CompanyListItem
     public function getName(): string
     {
         return $this->name;
+    }
+
+    /** @return null|string */
+    public function getCorporateName(): ?string
+    {
+        return $this->corporateName;
     }
 
     /**